Solved the timebug and added try/catch to avoid bug with missing info.
A bug caused the timer to count to zero and below on some occasions. It
was caused by a number error in the days calculation in the startTimer
function.

Also added try and catch to the variable assigning. This continues the
script even if some info is missing from the retrieved JSONP-object. It
also adds an error-message if throws an exception.